requirements.txt
# Core LLM / LangChain
langchain>=0.2.6,<0.3.0
langchain-community>=0.2.6,<0.3.0
langchain-core>=0.2.0,<0.3.0
langchain-openai>=0.1.0,<0.2.0
langchain-text-splitters>=0.2.0,<0.3.0
openai>=1.40.0,<2.0.0

# YouTube transcript extraction
youtube-transcript-api>=1.0.0,<2.0.0

# Gradio UI
gradio>=4.0.0,<5.0.0
huggingface_hub>=0.23.0,<0.25.0

# Environment / config
python-dotenv>=1.0.0

# YouTube channel listing
scrapetube>=2.5.0

# RAG: vector store + web scraping
faiss-cpu>=1.8.0
beautifulsoup4>=4.12.0

# Multi-agent framework
crewai[tools]>=1.0.0,<2.0.0

# Testing
pytest>=7.0.0
ruff>=0.4.0
